requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0

# Template
jinja2==3.1.3

# Auth
bcrypt==4.1.2
pyotp==2.9.0
python-jose[cryptography]==3.3.0

# Image
pillow==10.2.0

# Config
pyyaml==6.0.1
python-dotenv==1.0.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Utils
python-slugify==8.0.1
httpx==0.26.0
markupsafe==2.1.3

# Security
slowapi==0.1.9

# OAuth
authlib==1.3.0
